Move metadata parsing to separate object
